[Quasar] Consolidate SFPU tests into two shared kernel files
Replace 11 per-op C++ test files and 9 per-op Python test files with:
- sfpu_quasar_srca_test.cpp  (Unpack→SrcA→FPU MOVA2D→SFPU→Pack path)
- sfpu_quasar_dest_test.cpp  (Unpack→Dest→SFPU→Pack path)
- sfpu_quasar_ops.h          (shared op dispatcher, mirrors sfpu_operations.h)
- test_sfpu_quasar.py        (unified Python driver for all ops)

The Python selects which .cpp file to use per test via unpack_to_dest;
binary_mul_int is routed through the SrcA path to exercise FPU datacopy
for integer formats. Also adds the first Python-level test for
sfpu_binary_max_min. Also removes the redundant binary unpack configure
on the SrcA path (Quasar MOVA2D uses ELWADD+0 and does not require SrcB
to be configured).
