--- tt_metal/tt-llk/tests/python_tests/quasar/test_sfpu_rsqrt_quasar.py
@@ -1,192 +1,0 @@
-# SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-# SPDX-License-Identifier: Apache-2.0
-
-from typing import List
-
-import pytest
-import torch
-from helpers.format_config import DataFormat, FormatConfig
-from helpers.golden_generators import UnarySFPUGolden, get_golden_generator
-from helpers.llk_params import (
-    DataCopyType,
-    DestAccumulation,
-    DestSync,
-    ImpliedMathFormat,
-    MathOperation,
-    UnpackerEngine,
-    format_dict,
-)
-from helpers.param_config import (
-    input_output_formats,
-    is_invalid_quasar_sfpu_format_combination,
-    parametrize,
-)
-from helpers.stimuli_config import StimuliConfig
-from helpers.stimuli_generator import StimuliSpec, generate_stimuli
-from helpers.test_config import TestConfig
-from helpers.test_variant_parameters import (
-    DATA_COPY_TYPE,
-    DEST_INDEX,
-    DEST_SYNC,
-    IMPLIED_MATH_FORMAT,
-    MATH_OP,
-    NUM_FACES,
-    TEST_FACE_DIMS,
-    TILE_COUNT,
-    UNPACKER_ENGINE_SEL,
-)
-from helpers.utils import passed_test
-
-
-def generate_sfpu_rsqrt_combinations(
-    formats_list: List[FormatConfig],
-):
-    """
-    Generate SFPU rsqrt test combinations.
-
-    Args: Input-output format pairs
-
-    Returns: List of (format, dest_acc, dest_sync, implied_math_format, input_dimensions) tuples
-    """
-    combinations = []
-
-    dest_sync_modes = (DestSync.Half, DestSync.Full)
-    for fmt in formats_list:
-        in_fmt = fmt.input_format
-
-        dest_acc_modes = (
-            (DestAccumulation.Yes,)
-            if in_fmt.is_32_bit()
-            else (DestAccumulation.No, DestAccumulation.Yes)
-        )
-        for dest_acc in dest_acc_modes:
-            # Skip invalid format combinations for Quasar
-            if is_invalid_quasar_sfpu_format_combination(fmt, dest_acc):
-                continue
-
-            for dest_sync in dest_sync_modes:
-                for implied_math_format in [
-                    ImpliedMathFormat.No,
-                    ImpliedMathFormat.Yes,
-                ]:
-                    for input_dimensions in [[32, 32], [64, 64]]:
-                        combinations.append(
-                            (
-                                fmt,
-                                dest_acc,
-                                dest_sync,
-                                implied_math_format,
-                                input_dimensions,
-                            )
-                        )
-
-    return combinations
-
-
-SFPU_RSQRT_FORMATS = input_output_formats(
-    [
-        DataFormat.Float16,
-        DataFormat.Float32,
-        DataFormat.Float16_b,
-    ]
-)
-
-
-@pytest.mark.quasar
-@parametrize(
-    formats_dest_acc_sync_implied_math_input_dims=generate_sfpu_rsqrt_combinations(
-        SFPU_RSQRT_FORMATS
-    ),
-)
-def test_sfpu_rsqrt_quasar(formats_dest_acc_sync_implied_math_input_dims):
-    """
-    Test reciprocal square root (rsqrt) operation on Quasar architecture.
-
-    Uses PyTorch's rsqrt as the golden reference and generates input stimuli
-    covering the full representable range
-    """
-    (formats, dest_acc, dest_sync, implied_math_format, input_dimensions) = (
-        formats_dest_acc_sync_implied_math_input_dims[0]
-    )
-
-    sfpu_false_spec = StimuliSpec.uniform(low=0.0, high=1.0)
-    src_A, tile_cnt_A, src_B, _ = generate_stimuli(
-        stimuli_format_A=formats.input_format,
-        input_dimensions_A=input_dimensions,
-        stimuli_format_B=formats.input_format,
-        input_dimensions_B=input_dimensions,
-        spec_A=sfpu_false_spec,
-        spec_B=sfpu_false_spec,
-    )
-
-    # Scale to full representable range using log-uniform distribution
-    torch_format = format_dict[formats.input_format]
-    finfo = torch.finfo(torch_format)
-    min_val = max(1e-6, finfo.tiny * 100)
-    max_val = finfo.max
-
-    # Transform uniform [0,1) to log-uniform [min_val, max_val]
-    log_min = torch.log(torch.tensor(min_val, dtype=torch.float32))
-    log_max = torch.log(torch.tensor(float(max_val), dtype=torch.float32))
-    src_A = torch.exp(log_min + src_A.to(torch.float32) * (log_max - log_min)).to(
-        torch_format
-    )
-
-    num_faces = 4
-
-    generate_golden = get_golden_generator(UnarySFPUGolden)
-    golden_tensor = generate_golden(
-        MathOperation.Rsqrt,
-        src_A,
-        formats.output_format,
-        dest_acc,
-        formats.input_format,
-        input_dimensions,
-    )
-
-    configuration = TestConfig(
-        "sources/quasar/sfpu_rsqrt_quasar_test.cpp",
-        formats,
-        templates=[
-            MATH_OP(mathop=MathOperation.Rsqrt),
-            IMPLIED_MATH_FORMAT(implied_math_format),
-            DATA_COPY_TYPE(DataCopyType.A2D),
-            UNPACKER_ENGINE_SEL(UnpackerEngine.UnpA),
-            DEST_SYNC(dest_sync),
-        ],
-        runtimes=[
-            TILE_COUNT(tile_cnt_A),
-            NUM_FACES(num_faces),
-            TEST_FACE_DIMS(),
-            DEST_INDEX(0),
-        ],
-        variant_stimuli=StimuliConfig(
-            src_A,
-            formats.input_format,
-            src_B,
-            formats.input_format,
-            formats.output_format,
-            tile_count_A=tile_cnt_A,
-            tile_count_B=tile_cnt_A,
-            tile_count_res=tile_cnt_A,
-            num_faces=num_faces,
-        ),
-        unpack_to_dest=(
-            formats.input_format.is_32_bit() and dest_acc == DestAccumulation.Yes
-        ),
-        dest_acc=dest_acc,
-    )
-
-    res_from_L1 = configuration.run().result
-
-    # Verify results match golden
-    assert len(res_from_L1) == len(
-        golden_tensor
-    ), "Result tensor and golden tensor are not of the same length"
-
-    torch_format = format_dict[formats.output_format]
-    res_tensor = torch.tensor(res_from_L1, dtype=torch_format)
-
-    assert passed_test(
-        golden_tensor, res_tensor, formats.output_format
-    ), "Assert against golden failed"
