--- tt_metal/tt-llk/tests/sources/quasar/sfpu_binary_float_quasar_test.cpp
@@ -1,140 +1,0 @@
-// SPDX-FileCopyrightText: © 2026 Tenstorrent AI ULC
-//
-// SPDX-License-Identifier: Apache-2.0
-
-#include <cstdint>
-
-#include "ckernel.h"
-#include "llk_defs.h"
-#include "llk_memory_checks.h"
-#include "sfpu_stub.h"
-
-#ifdef LLK_TRISC_UNPACK
-
-#include "llk_math_common.h"
-#include "llk_unpack_common.h"
-#include "llk_unpack_unary_operand.h"
-#include "params.h"
-
-void run_kernel(RUNTIME_PARAMETERS params)
-{
-#if defined(RUNTIME_FORMATS) && !defined(SPEED_OF_LIGHT)
-    const FormatConfig& formats = params.formats;
-#endif
-    const std::uint32_t buf_desc_id          = 0;
-    const std::uint32_t num_tiles_per_unpack = params.TILE_CNT;
-
-    set_up_dest_dvalid_per_thread<dest_dvalid_client::UNPACK>({dest_dvalid_client::UNPACK, dest_dvalid_client::SFPU, dest_dvalid_client::PACK});
-    _llk_math_upk_to_dest_hw_configure_<IMPLIED_MATH_FORMAT, is_fp32_dest_acc_en, false /* EN_INT32_MATH_FORMAT */>();
-
-    buffer_descriptor_u bd_val = {0};
-
-    bd_val.f.l1_addr_16B = L1_ADDRESS(params.buffer_A[0]);
-    bd_val.f.format      = static_cast<std::uint8_t>(formats.unpack_A_src);
-    bd_val.f.x_dim       = params.TEST_FACE_C_DIM;
-    bd_val.f.y_dim       = params.TEST_FACE_R_DIM;
-    bd_val.f.z_dim       = params.num_faces;
-
-    tdma_descriptor_t td_val;
-    td_val.buf_desc        = bd_val;
-    td_val.buf_desc_id     = buf_desc_id;
-    td_val.reg_data_format = static_cast<std::uint8_t>(formats.unpack_A_dst);
-    _configure_buf_desc_table_(td_val.buf_desc_id, td_val.buf_desc);
-
-    _llk_unpack_configure_unary_<UNPACKER_ENGINE_SEL>(td_val);
-
-    _llk_unpack_unary_operand_init_<UNPACKER_ENGINE_SEL, false /* transpose */, is_fp32_dest_acc_en>(buf_desc_id, num_tiles_per_unpack);
-    _llk_unpack_unary_operand_<UNPACKER_ENGINE_SEL>(0);
-
-    _llk_unpack_dest_dvalid_section_done_<dest_sync>();
-}
-
-#endif
-
-#ifdef LLK_TRISC_MATH
-
-#include "cfg_defines.h"
-#include "cmath_common.h"
-#include "llk_math_common.h"
-#include "llk_math_eltwise_binary_sfpu.h"
-#include "llk_sfpu/ckernel_sfpu_binary.h"
-#include "params.h"
-
-using namespace ckernel;
-using namespace ckernel::math;
-using namespace ckernel::sfpu;
-
-void run_kernel(RUNTIME_PARAMETERS params)
-{
-#if defined(RUNTIME_FORMATS) && !defined(SPEED_OF_LIGHT)
-    const FormatConfig& formats = params.formats;
-#endif
-    set_up_dest_dvalid_per_thread<dest_dvalid_client::SFPU>({dest_dvalid_client::UNPACK, dest_dvalid_client::SFPU, dest_dvalid_client::PACK});
-
-    DataFormat src_format = static_cast<DataFormat>(formats.math);
-    _llk_math_srcAB_hw_configure_<IMPLIED_MATH_FORMAT, is_fp32_dest_acc_en, false /* EN_INT32_MATH_FORMAT */>(src_format, src_format);
-
-    // Mirrors the original L5 invocation: SFP_ROWS = 2 on Quasar, so a full face
-    // (TEST_FACE_R_DIM = 16 rows) corresponds to 8 sfpi row-pairs (SFPU_ITERATIONS).
-    // The iteration count is passed as a runtime arg, matching the L4 function signature.
-    const std::uint32_t num_sfpu_iterations = params.TEST_FACE_R_DIM / ckernel::math::SFP_ROWS;
-
-    _llk_math_eltwise_sfpu_init_();
-
-    // Programmable-constant init for the sfpi reciprocal helper. Sets
-    // `sfpi::vConstFloatPrgm0 = 2.0f` (the constant used by the in-helper
-    // Newton-Raphson refinement). No-op for MUL / when APPROXIMATION_MODE = true.
-    sfpu_binary_init<false /*APPROXIMATION_MODE*/, SFPU_BINARY_OPERATION>();
-
-    // sfpi vFloat binary op: op selected at compile time via BINOP if constexpr inside
-    // calculate_sfpu_binary; runtime iterations mirrors the original L5 calling convention.
-    _llk_math_eltwise_sfpu_params_(
-        calculate_sfpu_binary<false /*APPROXIMATION_MODE*/, SFPU_BINARY_OPERATION, is_fp32_dest_acc_en>,
-        0 /* DST base tile index */,
-        num_sfpu_iterations,
-        params.SRC0_TILE_IDX,
-        params.SRC1_TILE_IDX,
-        params.DST_TILE_IDX);
-
-    _llk_math_set_dvalid_<p_cleardvalid::SFPU, dest_sync>();
-}
-
-#endif
-
-#ifdef LLK_TRISC_PACK
-
-#include "cfg_defines.h"
-#include "llk_pack.h"
-#include "llk_pack_common.h"
-#include "params.h"
-
-void run_kernel(RUNTIME_PARAMETERS params)
-{
-#if defined(RUNTIME_FORMATS) && !defined(SPEED_OF_LIGHT)
-    const FormatConfig& formats = params.formats;
-#endif
-    const std::uint32_t buf_desc_id = 8;
-
-    set_up_dest_dvalid_per_thread<dest_dvalid_client::PACK>({dest_dvalid_client::UNPACK, dest_dvalid_client::SFPU, dest_dvalid_client::PACK});
-
-    buffer_descriptor_u bd_val = {0};
-    bd_val.f.l1_addr_16B       = L1_ADDRESS(params.buffer_Res[0]);
-    bd_val.f.format            = static_cast<std::uint8_t>(formats.pack_dst);
-    bd_val.f.x_dim             = params.TEST_FACE_C_DIM;
-    bd_val.f.y_dim             = params.TEST_FACE_R_DIM;
-    bd_val.f.z_dim             = params.num_faces;
-
-    tdma_descriptor_t tdma_desc;
-    tdma_desc.buf_desc        = bd_val;
-    tdma_desc.buf_desc_id     = buf_desc_id;
-    tdma_desc.reg_data_format = static_cast<std::uint8_t>(formats.pack_src);
-    _configure_buf_desc_table_(tdma_desc.buf_desc_id, tdma_desc.buf_desc);
-
-    _llk_pack_hw_configure_<p_pacr::PACK0>(tdma_desc);
-    _llk_pack_init_(buf_desc_id, 1);
-
-    _llk_pack_(params.DST_TILE_IDX, 0 /* tile index */);
-
-    _llk_pack_dest_dvalid_section_done_<dest_sync, is_fp32_dest_acc_en>();
-}
-#endif
